Merged from gioblu master

+++ src/strategies/ThroughSerial/ThroughSerial.h
@@ -178,12 +178,12 @@ class ThroughSerial {
 
     /* Send a string: */
 
-    void send_string(uint8_t *string, uint8_t length) {
+    void send_string(uint8_t *string, uint16_t length) {
       start_tx();
       uint16_t overhead = 2;
       // Add frame flag
       send_byte(TS_START);
-      for(uint8_t b = 0; b < length; b++) {
+      for(uint16_t b = 0; b < length; b++) {
         // Byte-stuffing
         if(
           (string[b] == TS_START) ||
